@@ -3,7 +3,7 @@ require 'spec_helper'
 module OpenStax::Accounts
   describe AnonymousAccount do
     it 'is anonymous' do
-      expect(AnonymousAccount.instance.is_anonymous?).to be_true
+      expect(AnonymousAccount.instance.is_anonymous?).to eq true
     end
   end
 end
